+++ lib/queue_bus/local.rb
@@ -6,7 +6,11 @@ module QueueBus
     class << self
       def publish(attributes = {})
         if ::QueueBus.local_mode == :suppress
-          ::QueueBus.log_worker("Suppressed: #{attributes.inspect}")
+          log_context = {
+            event_type: attributes['bus_event_type'],
+            bus_id: attributes['bus_id']
+          }
+          ::QueueBus.log_worker('Event suppressed in local mode', log_context)
           return # not doing anything
         end
 
@@ -14,7 +18,12 @@ module QueueBus
         json = ::QueueBus::Util.encode(attributes)
         attributes = ::QueueBus::Util.decode(json)
 
-        ::QueueBus.log_worker("Local running: #{attributes.inspect}")
+        log_context = {
+          event_type: attributes['bus_event_type'],
+          bus_id: attributes['bus_id'],
+          local_mode: ::QueueBus.local_mode
+        }
+        ::QueueBus.log_worker('Local mode processing event', log_context)
 
         # looking for subscriptions, not queues
         subscription_matches(attributes).each do |sub|